% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{climatch_perc}
\alias{climatch_perc}
\title{Percent Euclidean climatch score}
\usage{
climatch_perc(recipient, source, globvar)
}
\arguments{
\item{recipient}{List of data.frames of the recipient regions}

\item{source}{List of data.frames of the source regions}

\item{globvar}{Vector of the global variance of each variable}
}
\value{
Percent climatch scores within recipient >=6
}
\description{
Number of climatch scores >=6 in recipient region divided by number of rows of data (i.e., grid cells)
}
\examples{
i <- as.data.frame(matrix(runif(n=180, min=1, max=20), nrow=60)) # Fake source climate data.
j <- as.data.frame(matrix(runif(n=300, min=10, max=40), nrow=100)) # Fake recipient data.
variance <- c(600, 800, 450) # Fake global variance

climatch_perc(recipient = j, source = i, globvar = variance)
}
