% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{climatch_vec}
\alias{climatch_vec}
\title{Euclidean climatch scores}
\usage{
climatch_vec(recipient, source, globvar)
}
\arguments{
\item{recipient}{List of data.frames of the recipient regions}

\item{source}{List of data.frames of the source regions}

\item{globvar}{Vector of the global variance of each variable}
}
\value{
Vector of climatch scores corresponding to each grid cell within recipient
}
\description{
Vector of the climatch scores within the recipient region
}
\examples{
i <- as.data.frame(matrix(runif(n=180, min=1, max=20), nrow=60)) # Fake source climate data.
j <- as.data.frame(matrix(runif(n=300, min=10, max=40), nrow=100)) # Fake recipient data.
variance <- c(600, 800, 450) # Fake global variance

climatch_vec(recipient = j, source = i, globvar = variance)
}
